Functionality updated, Bill Details and Bill added and integrated to the project

+++ src/main/java/com/ggomezr/bookingsystem/domain/repository/RoomRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
     @Override
-    Optional<Room> findById(Long id);
+    Optional<Room> findById(Integer id);
 }
